refactor: group routing components
