+++ src/routing/RedirectIfLoggedInRoute.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Redirect, Route } from 'react-router-dom'
+
+// Import contexts
+const { LoginConsumer } = require('../contexts/LoginContext')
+
+export default function RedirectIfLoggedInRoute({ children, component: WrappedComponent, ...rest }) {
+  return (
+    <LoginConsumer>
+      { ({userId}) => (
+        userId ? (
+          <Redirect to="/sites" />
+        ) : (
+          children || WrappedComponent && <Route {...rest} component={WrappedComponent}/>
+        )
+      )}
+    </LoginConsumer>
+  )
+}
